Venda finalizada

+++ PetCafeProject/Models/Produto.cs
@@ -11,7 +11,7 @@ namespace PetCafeProject.Models
 
         public string? descricao { get; set; }
 
-        public float? valor { get; set; }
+        public double? valor { get; set; }
 
     }
